fix(profile): address pr comments
