+++ angular-ngrx-scss/scripts/generate-language-colors.js
@@ -1,5 +1,17 @@
 #!/usr/bin/env node
 
+/*
+ * The purpose of this script is to generate the language colors JSON file
+ * imported by the frontend. GitHub's REST API does not return language colors
+ * in its responses like the GraphQL API does.
+ *
+ * Thankfully GitHub provides colors in their linguist repository, but it has a
+ * lot of extra information and it isn't in an easy to digest format. This
+ * script downloads the latest version of that file and extracts the colors from
+ * it. Languages are keyed by name, which matches the language name in the API
+ * responses, and the value is the color hex code.
+ */
+
 import fetch from 'node-fetch';
 import fs from 'fs/promises';
 import yaml from 'yaml';
@@ -23,6 +35,4 @@ const main = async () => {
   await fs.writeFile(out, JSON.stringify(colors));
 };
 
-(async () => {
-  await main();
-})();
+await main();
