+++ angular-ngrx-scss/src/app/shared/pipes/relative-time.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { RelativeTimePipe } from './relative-time.pipe';
+
+describe('RelativeTimePipe', () => {
+  it('create an instance', () => {
+    const pipe = new RelativeTimePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
