+++ angular-ngrx-scss/src/app/shared/pipes/relative-time.pipe.ts
@@ -1,0 +1,11 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { relativeTimeFmt } from '../utils';
+
+@Pipe({
+  name: 'relativeTime',
+})
+export class RelativeTimePipe implements PipeTransform {
+  transform(value: string): unknown {
+    return relativeTimeFmt(value);
+  }
+}
