--- angular-ngrx-scss/src/app/state/profile/profile.actions.spec.ts
@@ -1,9 +1,0 @@
-import * as fromProfile from './profile.actions';
-
-describe('fetchProfiles', () => {
-  it('should return an action', () => {
-    expect(fromProfile.fetchProfile({ username: 'test' }).type).toBe(
-      '[Profile] Fetch Profile',
-    );
-  });
-});
