--- angular-ngrx-scss/src/app/state/profile/profile.reducer.spec.ts
@@ -1,13 +1,0 @@
-import { reducer, initialState } from './profile.reducer';
-
-describe('Profile Reducer', () => {
-  describe('an unknown action', () => {
-    it('should return the previous state', () => {
-      const action = {} as any;
-
-      const result = reducer(initialState, action);
-
-      expect(result).toBe(initialState);
-    });
-  });
-});
