--- angular-ngrx-scss/src/app/state/profile/profile.selectors.spec.ts
@@ -1,3 +1,0 @@
-describe('Profile Selectors', () => {
-  // it('should select the feature state', () => {});
-});
