+++ angular-ngrx-scss/src/app/state/user/user.effects.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { provideMockActions } from '@ngrx/effects/testing';
 import { Action } from '@ngrx/store';
 import { Observable, of } from 'rxjs';
-import { fetchUserData, fetchUserDataSuccess } from '.';
+import { fetchUserData, fetchUserDataSuccess } from './user.actions';
 import { UserService } from '../../user/services/user.service';
 import { UserEffects } from './user.effects';
 import { UserState } from './user.state';
