+++ angular-ngrx-scss/tsconfig.json
@@ -1,7 +1,5 @@
 /* To learn more about this file see: https://angular.io/config/tsconfig. */
 {
-  "include": ["src/**/*.ts", "typings.d.ts"],
-  "exclude": ["src/**/*.spec.ts"],
   "compileOnSave": false,
   "compilerOptions": {
     "baseUrl": "./",
@@ -22,7 +20,6 @@
     "target": "es2017",
     "module": "es2020",
     "lib": ["es2020", "dom"],
-    "typeRoots": ["node_modules/@types", "typings.d.ts"],
     "allowSyntheticDefaultImports": true
   },
   "angularCompilerOptions": {
